Update and rename content/authentication/index.md to içerik/kimlik doğrulama/index.md Turkey
One of the most effective measures that can be taken against the cyber attacks of the new era is to add a second or third verification (two-factor authentication) such as a pin, one-time codes (one time passcodes) or biometrics, in addition to a single password. is to switch to the strong authentication) method. This has now turned into an indispensable requirement, especially in corporate infrastructures.
In many of the major data leaks in recent years, compromised user logins have played a large part. The security of systems, applications and platforms connected with only a single user/password is until that password is exposed and compromised. Since emerging threats are often able to break through a single password barrier, strong authentication is now required by making a second and sometimes even a third authentication instead of a single password.